Changes following code review
